custom permissions, IsOwnerOrReadOnly class

+++ drf/home/views.py
@@ -3,12 +3,13 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from .models import Person, Question, Answer
 from .serializers import PersonSerializer, QuestionSerializer, AnswerSerializer
-from rest_framework.permissions import IsAuthenticated, IsAdminUser
+from rest_framework.permissions import IsAuthenticated, IsAdminUser, AllowAny
 from rest_framework import status
-
+from permissions import IsOwnerOrReadOnly
 
 class Home(APIView):
-    permission_classes = [IsAuthenticated,]
+    
+    permission_classes = [AllowAny,]
     
     def get(self, request):
         persons = Person.objects.all()
@@ -18,6 +19,8 @@ class Home(APIView):
 
 class QuestionListView(APIView):
     
+    permission_classes = [IsAuthenticated,]
+    
     def get(self, request):
         questions = Question.objects.all()
         srz_data = QuestionSerializer(instance=questions, many=True).data
@@ -25,6 +28,8 @@ class QuestionListView(APIView):
 
     
 class QuestionCreateView(APIView):
+    
+    permission_classes = [IsAuthenticated,]
     
     def post(self, request):
         srz_data = QuestionSerializer(data=request.POST)
@@ -37,9 +42,12 @@ class QuestionCreateView(APIView):
     
 class QuestionUpdateView(APIView):
     
+    permission_classes = [IsOwnerOrReadOnly,]
+    
     def put(self, request, pk):
         question = Question.objects.get(pk=pk)
         srz_data = QuestionSerializer(instance=question, data=request.data, partial=True)
+        self.check_object_permissions(request, question)
         if srz_data.is_valid():
             srz_data.save()
             return Response(srz_data.data, status=status.HTTP_200_OK)
@@ -48,8 +56,11 @@ class QuestionUpdateView(APIView):
     
 class QuestionDeleteView(APIView):
     
+    permission_classes = [IsOwnerOrReadOnly,]
+    
     def delete(self, request, pk):
         question = Question.objects.get(pk=pk)
+        self.check_object_permissions(request, question)
         question.delete()
         return Response({'message': 'deleted successfully'}, status=status.HTTP_200_OK)
